moved require32ByteSharedSecrets to EnabledFeatureSettings
