+++ connector-server/src/main/resources/application.yml
@@ -97,7 +97,6 @@ interledger:
       # Base64-encoded String of bytes used to derive any values used in encryption/decryption for SPSP.
       # EXAMPLE: serverSecret: aQLR5IWAGV2vKnBhnFFsl2cXOCh9u0IWz3PiA64KlJ8=
       addressPrefixSegment: spsp
-    require32ByteSharedSecrets: true
     keys:
       secret0:
         alias: secret0
@@ -108,6 +107,7 @@ interledger:
     enabledFeatures:
       rateLimitingEnabled: false
       localSpspFulfillmentEnabled: false
+      require32ByteSharedSecrets: true
     enabledProtocols:
       ilpOverHttpEnabled: true
       pingProtocolEnabled: true
