+++ connector-service-impl/src/test/java/org/interledger/connector/links/DefaultLinkSettingsValidatorTest.java
@@ -3,12 +3,14 @@ package org.interledger.connector.links;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 import org.interledger.connector.crypto.ConnectorEncryptionService;
 import org.interledger.connector.settings.ConnectorSettings;
+import org.interledger.connector.settings.EnabledFeatureSettings;
 import org.interledger.crypto.Decryptor;
 import org.interledger.crypto.EncryptedSecret;
 import org.interledger.link.ImmutableLinkSettings;
@@ -60,7 +62,9 @@ public class DefaultLinkSettingsValidatorTest {
   public void setUp() {
     initMocks(this);
 
-    when(connectorSettings.isRequire32ByteSharedSecrets()).thenReturn(false);
+    EnabledFeatureSettings enabledFeatureMock = mock(EnabledFeatureSettings.class);
+    when(connectorSettings.enabledFeatures()).thenReturn(enabledFeatureMock);
+    when(enabledFeatureMock.isRequire32ByteSharedSecrets()).thenReturn(false);
     when(mockConnectorEncryptionService.getDecryptor()).thenReturn(decryptor);
 
     when(decryptor.withDecrypted(eq(ENCRYPTED_INCOMING_SECRET), any()))
@@ -114,7 +118,7 @@ public class DefaultLinkSettingsValidatorTest {
 
   @Test
   public void require32ByteSecretFailsIfConfigured() {
-    when(connectorSettings.isRequire32ByteSharedSecrets()).thenReturn(true);
+    when(connectorSettings.enabledFeatures().isRequire32ByteSharedSecrets()).thenReturn(true);
 
     final IlpOverHttpLinkSettings linkSettings = newSettings(INCOMING_BASE_64, OUTGOING_BASE_64);
